1-29-21 Mapping Update
The script crashed on characters for some reason so this map will take a shile to load because every character needs to be randomly generated. Not sure if that's the reason why there are a bunch of random provinces with incorrect assignments. Any help identifying the common denominator would be helpful.